Add better exiting

+++ run_server.py
@@ -3,12 +3,14 @@ import re
 import signal as signals
 import sys
 from asyncio.streams import StreamReader
+from contextlib import asynccontextmanager
 from dataclasses import dataclass
 from os import environ as env
 from pathlib import Path
-from typing import Optional, TextIO
+from typing import AsyncGenerator, AsyncIterator, Optional, TextIO
 
 import a2s
+from asyncio_mqtt import Client as MQTTClient
 
 THIS_DIR = Path(__file__).parent.absolute()
 LD_LIBRARY_PATH = str(THIS_DIR.joinpath("vhserver/linux64"))
@@ -37,26 +39,51 @@ class ValheimServer:
         self.sigints = 0
 
         self.started = asyncio.Event()
+        self.stopping = asyncio.Event()
         self.finished = asyncio.Event()
 
         self.stdout_task: "Optional[asyncio.Task[None]]" = None
         self.stderr_task: "Optional[asyncio.Task[None]]" = None
         self.status_task: "Optional[asyncio.Task[None]]" = None
+        self.stop_task: "Optional[asyncio.Task[None]]" = None
+
+        self.log_line_queue: "Optional[asyncio.Queue[str]]" = None
 
         self.log_triggers = {
             r".+Done generating locations.+": self.handle_locations_generated
         }
 
         def sigint_handler() -> None:
-            self.stop()
-            if self.proc is not None:
-                if self.sigints >= 2:
-                    self.proc.kill()
+            self.sigints += 1
+            if not self.stopping.is_set():
+                self.stop_task = self.loop.create_task(self.stop())
+            if self.proc is not None and self.sigints >= 2:
+                print("***Killing Server***")
+                self.proc.kill()
+                self.finished.set()
 
         self.loop.add_signal_handler(signals.SIGINT, sigint_handler)
 
     def handle_locations_generated(self, match):
         self.started.set()
+
+    @asynccontextmanager
+    async def log_lines(self) -> AsyncIterator[AsyncGenerator[str, None]]:
+        self.log_line_queue = asyncio.Queue()
+
+        async def log_generator() -> AsyncGenerator[str, None]:
+            finished = self.loop.create_task(self.finished.wait())
+            while not self.finished.is_set():
+                log_line = self.loop.create_task(self.log_line_queue.get())
+                done, _ = await asyncio.wait(
+                    (finished, log_line), return_when=asyncio.FIRST_COMPLETED
+                )
+                if finished in done:
+                    return
+                yield await log_line
+
+        yield log_generator
+        self.log_line_queue = None
 
     async def __aenter__(self):
         args = [
@@ -85,14 +112,23 @@ class ValheimServer:
         self.stderr_task = self.loop.create_task(self.print_stderr())
         self.status_task = self.loop.create_task(self.print_status())
 
-        return self
+        return self.log_lines()
 
     async def __aexit__(self, exc_type, exc, tb):
+        if not self.stopping.is_set():
+            await self.stop()
+
+    async def stop(self):
         print("*** Stopping the server ***")
-        if self.proc is not None:
+        self.stopping.set()
+        if self.proc is not None and self.proc.returncode is None:
+            # Maybe just try catching the case in which the process doesn't exist,
+            # because it could quit at any moment, even though returncode is not None.
             self.proc.send_signal(signals.SIGINT)
             exit_code = await self.proc.wait()
             print(f"Server exited with code {exit_code}")
+
+        self.finished.set()
 
         self.stdout_task.cancel()
         self.stderr_task.cancel()
@@ -107,11 +143,13 @@ class ValheimServer:
         while True:
             line = (await input.readline()).decode("utf8")
             if not line:
-                await asyncio.sleep(1)
-                continue
+                # EOF so the pipe has been closed
+                return
             if line.strip() in IGNORE_OUTPUT:
                 continue
             print(line, file=output, end="")
+            if self.log_line_queue is not None:
+                self.log_line_queue.put_nowait(line)
             for pattern, handler in self.log_triggers.items():
                 match = re.match(pattern, line)
                 if match:
@@ -126,12 +164,14 @@ class ValheimServer:
     async def print_status(self) -> None:
         address = ("localhost", self.port + 1)
         await self.started.wait()
-        while True:
+        while not self.stopping.is_set():
             try:
                 info, players = await asyncio.gather(
                     a2s.ainfo(address),
                     a2s.aplayers(address),
                 )
+            except asyncio.CancelledError:
+                raise
             except Exception as exc:
                 print(f"Exception getting status: {type(exc)}: {exc}")
                 await asyncio.sleep(5)
@@ -144,7 +184,9 @@ class ValheimServer:
 def main():
     async def async_main():
         async with ValheimServer("Test", "noworld", "testingzzzzzzzz") as vhs:
-            await vhs.started.wait()
+            async with vhs as log_lines:
+                async for line in log_lines():
+                    print("oooo")
 
     asyncio.run(async_main())
